Sub and Pub with MQTT works + saving to firestore

+++ SPRING REST API/src/main/java/no/hvl/dat250/feedapp/repository/PollRepository.java
@@ -15,6 +15,6 @@ public interface PollRepository extends CrudRepository<Poll, Long> {
 	Iterable<Poll> findByTitleContaining(String title);
 	
 	@Query("select id from Poll p where p.endDate < CURRENT_TIMESTAMP and p.endDate > :prev")//TODO: DETTE ER BROKEN 
-	Optional<List<Integer>> findEndedPolls(@Param("prev")Timestamp prev);
+	Optional<List<Long>> findEndedPolls(@Param("prev")Timestamp prev);
 
 }
